requirements.txt
playwright
python-dotenv
beautifulsoup4
pdfplumber
